+++ code/modules/reagents/reagents/toxins.dm
@@ -72,6 +72,8 @@
 	heating_point = 523
 	heating_products = list("toxin")
 	reagent_type = "Toxin/Stimulator"
+	withdrawal_threshold = 10
+	withdrawal_rate = REM
 
 /datum/reagent/toxin/carpotoxin/affect_blood(mob/living/carbon/M, alien, effect_multiplier)
 	if(ishuman(M))
@@ -79,7 +81,7 @@
 		var/obj/item/organ/internal/liver/L = H.random_organ_by_process(OP_LIVER)
 		if(istype(L))
 			L.take_damage(strength * effect_multiplier, 0)
-	M.stats.addTempStat(STAT_VIG, STAT_LEVEL_BASIC, STIM_TIME, "carpotoxin")
+	M.stats.addTempStat(STAT_VIG, STAT_LEVEL_ADEPT, STIM_TIME, "carpotoxin")
 
 /datum/reagent/toxin/carpotoxin/withdrawal_act(mob/living/carbon/M)
 	M.stats.addTempStat(STAT_VIG, -STAT_LEVEL_BASIC, STIM_TIME, "carpotoxin_w")
@@ -251,7 +253,7 @@
 	return TRUE
 
 /datum/reagent/toxin/plantbgone/touch_obj(obj/O, var/volume)
-	if(istype(O, /obj/effect/plant))
+	if(istype(O, /obj/effect/plant) && !istype(O, /obj/effect/plant/hivemind))
 		qdel(O)
 	if(istype(O, /obj/machinery/portable_atmospherics/hydroponics))
 		var/obj/machinery/portable_atmospherics/hydroponics/T = O
@@ -439,7 +441,7 @@
 	ADD_TRANSFORMATION_MOVEMENT_HANDLER(M)
 	M.canmove = 0
 	M.icon = null
-	M.cut_overlays()
+	M.overlays.Cut()
 	M.invisibility = 101
 	for(var/obj/item/W in M)
 		if(istype(W, /obj/item/implant)) //TODO: Carn. give implants a dropped() or something
@@ -535,6 +537,7 @@
 	heating_point = 573
 	heating_products = list("radium", "acetone", "hydrazine", "nutriment")
 	reagent_type = "Toxin/Stimulator"
+	withdrawal_threshold = 15
 
 /datum/reagent/toxin/diplopterum/affect_blood(mob/living/carbon/M, alien, effect_multiplier)
 	..()
@@ -567,6 +570,7 @@
 	nerve_system_accumulations = 5
 	heating_point = 573
 	heating_products = list("radium", "ammonia", "sulfur", "nutriment")
+	withdrawal_threshold = 10
 
 /datum/reagent/toxin/seligitillin/affect_blood(mob/living/carbon/M, alien, effect_multiplier)
 	M.add_chemical_effect(CE_BLOODCLOT, 0.2)
@@ -604,6 +608,7 @@
 	heating_point = 573
 	heating_products = list("radium", "aluminum", "tungsten", "nutriment")
 	reagent_type = "Toxin/Stimulator"
+	withdrawal_threshold = 10
 
 /datum/reagent/toxin/starkellin/affect_blood(mob/living/carbon/M, alien, effect_multiplier)
 	..()
@@ -627,6 +632,7 @@
 	heating_point = 573
 	heating_products = list("radium", "mercury", "sugar", "nutriment")
 	reagent_type = "Toxin/Stimulator"
+	withdrawal_threshold = 10
 
 /datum/reagent/toxin/gewaltine/affect_blood(mob/living/carbon/M, alien, effect_multiplier)
 	..()
@@ -652,8 +658,10 @@
 	nerve_system_accumulations = 4
 	heating_point = 573
 	heating_products = list("radium", "mercury", "lithium", "nutriment")
+	withdrawal_threshold = 5
 
 /datum/reagent/toxin/fuhrerole/affect_blood(mob/living/carbon/M, alien, effect_multiplier)
+	M.stats.addTempStat(STAT_VIG, STAT_LEVEL_ADEPT * effect_multiplier, STIM_TIME, "fuhrerole_w")
 	M.faction = "roach"
 
 /datum/reagent/toxin/fuhrerole/on_mob_delete(mob/living/L)
@@ -682,12 +690,13 @@
 	heating_point = 573
 	heating_products = list("fuhrerole", "radium", "nutriment", "tungsten")
 	reagent_type = "Toxin/Stimulator"
+	withdrawal_threshold = 10
 
 /datum/reagent/toxin/kaiseraurum/affect_blood(mob/living/carbon/M, alien, effect_multiplier)
 	..()
 	M.stats.addTempStat(STAT_VIG, STAT_LEVEL_ADEPT * effect_multiplier, STIM_TIME, "kaiseraurum")
-	M.stats.addTempStat(STAT_TGH, STAT_LEVEL_ADEPT * effect_multiplier, STIM_TIME, "kaiseraurum")
-	M.stats.addTempStat(STAT_ROB, STAT_LEVEL_ADEPT * effect_multiplier, STIM_TIME, "kaiseraurum")
+	M.stats.addTempStat(STAT_TGH, STAT_LEVEL_BASIC * effect_multiplier, STIM_TIME, "kaiseraurum")
+	M.stats.addTempStat(STAT_ROB, STAT_LEVEL_BASIC * effect_multiplier, STIM_TIME, "kaiseraurum")
 	M.faction = "roach"
 
 /datum/reagent/toxin/kaiseraurum/withdrawal_act(mob/living/carbon/M)
@@ -741,6 +750,10 @@
 	taste_description = "bleach"
 	color = "#707c13"
 	strength = 15
+/datum/reagent/toxin/chlorine/touch_obj(obj/O)
+	if(istype(O, /obj/effect/plant/hivemind))
+		qdel(O)
+
 
 /datum/reagent/toxin/tar
 	name = "Tar"
@@ -749,3 +762,21 @@
 	color = "#140b30"
 	reagent_state = LIQUID
 	strength = 4
+
+/datum/reagent/toxin/mold
+	name = "Mold"
+	id = "mold"
+	description = "Food that went rotting for so long it liquefied. Do not consume."
+	taste_description = "stale vomit mixed with pineapples"
+	reagent_state = LIQUID
+	color = "#467508"
+	metabolism = REM
+	overdose = REAGENTS_OVERDOSE/2
+	nerve_system_accumulations = 5
+	strength = 0.01
+	sanityloss = 2
+
+/datum/reagent/toxin/mold/affect_ingest(mob/living/carbon/M, alien, effect_multiplier)
+	. = ..()
+	if(prob(20 * effect_multiplier))
+		M.vomit()
